Backward and forward pass can now use different time steps
I modified the code for the backwards pass so that it can use a different time step size than the forward pass. This works using an interpolation. I think this approach may be helpful in the future.

I also fixed a small bug in the solver (the final time was wrong).

In my next commit, I plan to remove the train function (move it back into the notebook), along with interpolate (also move it back into the notebook)

From there, I plan to add a RK solver.