Added integral loss adjoint class
This forced me to write a different class, since the other loss function requires a different adjoint/gradient expression.

The new function takes l and the true trajectory (or, rather, an interpolation of it) as arguments. I have not debugged the new code... I just got it running.

I had to modify the loss function to make this work. See loss.py.