Fixed general case in code.
The equations in the paper were incorrect. There should have been a - in front of the dL/dtheta equation. Likewise there should not be a - in front of the dL/dtau equation. I fixed both of these mistakes (in the paper, and in the code). With these changes, everything works!