Started major code enhancements
My plan is to make this code work for a general loss function, when the trajectory has an arbitrary number of variables, for an arbitrary right hand side. Making this work will take some time.

In this commit, I made the code run with the full loss function.

In the next commit, I plan to revise the code so that it can run when the right hand side has an arbitrary number of dimensions. To do this, I need to change computing the gradients to computing gradient vector products.

After that, I think I know how to make it work with an arbitrary right hand side.