I added the debug plots back into NDDE
They now print components of the jacobian vector product, however.

In my next commit, I want to consider enabling RK2.... or at least allow that to be a thing.

I also want to generaize the code to work with an arbitrary list of right hand side tensors... this should be fairly straight forward, I just need to do it.