Added support for arbitrary right-hand sides.
The right hand side of the DDE can now be an arbitrary torch.nn.Module object.

To make this work, I had to stop tracking the (dFdTheta)(t)^T p(t). I could have kept this, but the memory requirements would have grown too large. This forced me to compute the integral for the gradeints in the main loop (where we compute p). This made the code a little ugly, but it works...