X0 is now a torch.nn.Module object
I derived an expression for the gradeint of the loss with respect to parameters in the IC. This allows me to parameterize the IC as an arbitrary function, not just a learnable constant.

Making this change was... tricky. I had to modify a lottttt of the code to expect X0 as a module, not a fixed constant. Regardless, I was able to do it (after a lot of debugging....)

The code appears to work, too.

When I reduce X0 to a constant function, it works exactly as before (converges in the same number of steps)

I then wrote an IC that uses an affine map (t -> a*t + b). It can learn a and b!!!!!

I created a new file, X0, which houses the class definitions for the initial condition classes. Currently, the code is set up to learn affine initial conditions.