Switched from columns to rows for time axis
I should have made this change a very very long time ago. Most modern libraries expect the 0 axis to correspond to time, not the -1 axis. For some reason, when Maria first started this project, she choose the latter. When I took over, I continued using her approach. However, I wanted to replace the interpolation function (interp1d is outdated/legacy code). The problem is that modern scipy interpolators return a trajctory where the 0 axis is for time, not the -1 one. This made for some awkward conflicts. To fix it, I switched interp1d to CubicSpline and switched to using the 0 axis for time. This requird tweaking alot of stuff... but it appears to have worked.