Added noise, fixed a small bug
I added a function that can add noise to trajectories.

I also found a small bug... before, the solver broke [0, T] into increments of size tau/10. However, it set the time values of the solve to 0, (T/N), 2(T/N), ... The problem with this is that T/N is NOT dt... at least not in general. This meant that going back N time steps is not the same thing as going back tau in time.

to fix this, I now solve to N*dt, where N = min{ N : N*dt >= T}. This means I go beyond the final time, but this is fine... since I interpolate the solution anyway to compute the gradients.

I need to make a similar fix to the euler solver (I only did RK right now).

I also need to make a fix to the backward method, because it does the same thing. I put a temporary fix in place... but need to put a perminant one in at some point.